--- a/nim/config.nims
+++ b/nim/config.nims
@@ -1,16 +1,18 @@
 switch("opt", "speed")
-switch("cc", "clang")
+switch("cc", "gcc")
 switch("panics", "off")
-switch("passC", "-flto")
-switch("passL", "-flto")
+switch("passC", "-s -flto")
+switch("passL", "-s -flto")
 
 when defined(profileGen):
+  switch("cc", "clang")
   echo "Build with profileGen"
   switch("passC", "-fprofile-instr-generate")
   switch("passL", "-fprofile-instr-generate")
 
 when defined(profileUse):
   echo "Build with profileUse"
+  switch("cc", "clang")
   switch("passC", "-fprofile-instr-use")
   switch("passL", "-fprofile-instr-use")