@michaelsbradleyjr 2023-10-19
@inv2004 I'm getting pretty decent numbers with `gcc` PGO when generating the profile data with 60k posts.

See: https://github.com/jinyus/related_post_gen/pull/339#issuecomment-1771851138

Maybe checkout that branch (with latest commits) and see if you get similarly improved numbers.