@inv2004 2023-10-20
@michaelsbradleyjr 
It is good, but, unfortunately I think it is more like blindly attempts to guess how the test works. It does not match any my attempts to emulate it on ubuntu-22 and arch

That I definitely saw locally that old clang+PGO is visible worse than fresh gcc. Also clang+PGO generates pretty random build - it could fast or it could be 2x slower than prev build

You can find my post here (via translator): https://translated.turbopages.org/proxy_u/ru-en.ru.fe85e5f0-6531c4ec-9b8ea3e4-74722d776562/https/t.me/s/inv2004_dev_blog/97