@michaelsbradleyjr 2023-10-20
> It is good, but, unfortunately I think it is more like blindly attempts to guess how the test works. It does not match any my attempts to emulate it on ubuntu-22 or arch

The `archlinux:base` container made with this project's Dockerfile has `gcc-13.2` and `clang-16`. In #339 the latest numbers (my most recent comment at bottom) are from `gcc-13.1` on Ubuntu 22.04.

On my macOS laptop, `clang-15` PGO profiled on 60k shows a good speedup for multi-threaded, not so much for single-threaded.

I've done `clang-16` PGO on a local `archlinux:base` container with 5k posts and the speedup for single and multi is inconsistent and not very good, same as observed previously. I can try `clang-16` (or newer) PGO with 60k posts, and it would probably be decent, but since `gcc-13` profiled on 60k is working well, I hadn't tried it yet.