@michaelsbradleyjr 2023-10-20
I tried `clang-16` PGO profiled on 60k posts on Ubuntu 22.04. Consistent outcome for several builds/runs: harmful for single-threaded and mediocre for multi-threaded compared to `gcc-13` PGO 60k.

In my Azure VM, I made these changes on `nim_rev_5` branch
```diff
diff --git a/nim/config.nims b/nim/config.nims
index 66f99c8..92e198a 100644
--- a/nim/config.nims
+++ b/nim/config.nims
@@ -6,6 +6,7 @@ const
   cacheSubdir = joinPath(cacheSubdirHead,
     (if defined(release): "release" else: "debug"), projectName())
 
+switch("cc", "clang")
 switch("nimcache", cacheSubdir)
 switch("opt", "speed")
 switch("panics", "off")
diff --git a/nim/related.nimble b/nim/related.nimble
index 4b4ee23..2a8de91 100644
--- a/nim/related.nimble
+++ b/nim/related.nimble
@@ -24,9 +24,9 @@ task buildopt, "build optimized":
   exec "cd .. && python gen_fake_posts.py 60000"
   exec "./related >/dev/null"
   exec "cd .. && mv posts_orig.json posts.json"
-  when defined(clang):
-    when defined(macosx):
-      exec "xcrun llvm-profdata merge default*.profraw --output default.profdata"
-    else:
-      exec "llvm-profdata merge default*.profraw --output default.profdata"
+  # when defined(clang):
+  #   when defined(macosx):
+  #     exec "xcrun llvm-profdata merge default*.profraw --output default.profdata"
+  #   else:
+  exec "llvm-profdata merge default*.profraw --output default.profdata"
   exec "nimble --verbose build -d:profileUse -d:release"
diff --git a/nim_con/buildopt.sh b/nim_con/buildopt.sh
index 23d2b5b..44f1a8e 100755
--- a/nim_con/buildopt.sh
+++ b/nim_con/buildopt.sh
@@ -41,13 +41,13 @@ cd nim_con
 cd ..
 mv posts_orig.json posts.json
 cd nim_con
-if [[ "$(cc --version 2>/dev/null | head -1)" = *"clang"* ]]; then
-    if [[ -n "$(which xcrun 2>/dev/null)" ]]; then
-        xcrun llvm-profdata merge default*.profraw --output default.profdata
-    else
-        llvm-profdata merge default*.profraw --output default.profdata
-    fi
-fi
+# if [[ "$(cc --version 2>/dev/null | head -1)" = *"clang"* ]]; then
+#     if [[ -n "$(which xcrun 2>/dev/null)" ]]; then
+#         xcrun llvm-profdata merge default*.profraw --output default.profdata
+#     else
+llvm-profdata merge default*.profraw --output default.profdata
+#     fi
+# fi
 
 echo "Compiling optimized executable"
 nim c -d:FixedChanSize=${chansize} \
diff --git a/nim_con/config.nims b/nim_con/config.nims
index b55e221..dcdf05e 100644
--- a/nim_con/config.nims
+++ b/nim_con/config.nims
@@ -8,6 +8,7 @@ const
 
 switch("nimcache", cacheSubdir)
 
+--cc:clang
 --mm:arc
 --outdir:build
 --tlsEmulation:off # default on|off varies by platform
```
Results
```
(related) michael@azure1:~/repos/related_post_gen$ ./nim.sh 
  main
* nim_rev_5
Running Nim
...
gcc (Ubuntu 13.1.0-8ubuntu1~22.04) 13.1.0
Ubuntu clang version 16.0.6 (++20230710042027+7cbf1a259152-1~exp1~20230710162048.105)
...
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 51.222ms
	Processing time (w/o IO): 50.956ms
	Processing time (w/o IO): 50.93ms
	Processing time (w/o IO): 51.25ms
	Processing time (w/o IO): 51.003ms
	Processing time (w/o IO): 50.84ms
	Processing time (w/o IO): 51.011ms
	Processing time (w/o IO): 51.116ms
	Processing time (w/o IO): 50.888ms
	Processing time (w/o IO): 50.946ms
	Processing time (w/o IO): 50.961ms
	Processing time (w/o IO): 51.037ms
	Processing time (w/o IO): 50.823ms
	 Time (mean ± σ):      79.4 ms ±   0.4 ms    [User: 70.4 ms, System: 9.0 ms]
	 Range (min … max):    78.9 ms …  80.2 ms    10 runs
	
Checking output
Success: related_posts_nim.json is valid!
Running Nim
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 757.578ms
	Processing time (w/o IO): 760.062ms
	Processing time (w/o IO): 760.36ms
	 Time (mean ± σ):     878.9 ms ±   3.0 ms    [User: 852.8 ms, System: 25.9 ms]
	 Range (min … max):   876.8 ms … 881.1 ms    2 runs
	
Running Nim
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 6748.846ms
	Processing time (w/o IO): 6758.8ms
	Processing time (w/o IO): 6752.059ms
	 Time (mean ± σ):      7.150 s ±  0.001 s    [User: 7.061 s, System: 0.080 s]
	 Range (min … max):    7.149 s …  7.150 s    2 runs
```
```
(related) michael@azure1:~/repos/related_post_gen$ ./nim_con.sh 
  main
* nim_rev_5
Running Nim Concurrent
using 4 threads
...
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 24.249ms
	Processing time (w/o IO): 24.341ms
	Processing time (w/o IO): 24.57ms
	Processing time (w/o IO): 24.498ms
	Processing time (w/o IO): 24.166ms
	Processing time (w/o IO): 24.443ms
	Processing time (w/o IO): 25.375ms
	Processing time (w/o IO): 25.915ms
	Processing time (w/o IO): 24.727ms
	Processing time (w/o IO): 24.915ms
	Processing time (w/o IO): 24.812ms
	Processing time (w/o IO): 24.003ms
	Processing time (w/o IO): 24.356ms
	 Time (mean ± σ):      53.7 ms ±   1.4 ms    [User: 80.7 ms, System: 9.9 ms]
	 Range (min … max):    52.0 ms …  56.2 ms    10 runs
	
Checking output
Success: related_posts_nim_con.json is valid!
Running Nim Concurrent
using 4 threads
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 339.27ms
	Processing time (w/o IO): 340.43ms
	Processing time (w/o IO): 340.865ms
	 Time (mean ± σ):     457.9 ms ±   1.0 ms    [User: 990.3 ms, System: 26.0 ms]
	 Range (min … max):   457.2 ms … 458.6 ms    2 runs
	
Running Nim Concurrent
using 4 threads
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 3043.759ms
	Processing time (w/o IO): 3040.96ms
	Processing time (w/o IO): 3061.098ms
	 Time (mean ± σ):      3.440 s ±  0.016 s    [User: 8.490 s, System: 0.094 s]
	 Range (min … max):    3.429 s …  3.451 s    2 runs
```