@jinyus 2023-10-20
Main:
```
Nim | 30.41 ms | 434.66 ms | 3.82 s
```

This PR:
```
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 25.239ms
        Processing time (w/o IO): 25.01ms
        Processing time (w/o IO): 24.983ms
        Processing time (w/o IO): 24.923ms
        Processing time (w/o IO): 24.923ms
        Processing time (w/o IO): 25.12ms
        Processing time (w/o IO): 24.907ms
        Processing time (w/o IO): 24.955ms
        Processing time (w/o IO): 25.071ms
        Processing time (w/o IO): 25.179ms
        Processing time (w/o IO): 24.941ms
        Processing time (w/o IO): 24.862ms
        Processing time (w/o IO): 24.952ms
          Time (mean ± σ):      49.7 ms ±   0.9 ms    [User: 38.1 ms, System: 11.6 ms]
          Range (min … max):    48.8 ms …  51.2 ms    10 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 351.762ms
        Processing time (w/o IO): 351.961ms
        Processing time (w/o IO): 352.405ms
          Time (mean ± σ):     449.4 ms ±   0.3 ms    [User: 417.9 ms, System: 31.4 ms]
          Range (min … max):   449.2 ms … 449.6 ms    2 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 3099.301ms
        Processing time (w/o IO): 3091.127ms
        Processing time (w/o IO): 3088.197ms
          Time (mean ± σ):      3.443 s ±  0.009 s    [User: 3.333 s, System: 0.102 s]
          Range (min … max):    3.437 s …  3.449 s    2 runs
```